requirements.txt
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
SQLAlchemy==2.0.17
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
